Improvements to contrastCheck function
Use Compass helper pow() to raise the expressions that calculate the
R, G and B components by the power of 2.4 (as per W3C equation for
relative luminance).
Use lightness() function to brigten a foreground color in the same
color range to increase the contrast.

TODO: if brightening the color *still* doesn't achieve enough contrast
then darken the color until it does.